requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Authentication and Security
PyJWT>=2.8.0
cryptography>=46.0.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
hypothesis>=6.92.1

# AI and ML
google-generativeai>=0.3.2
openai>=1.3.7
anthropic>=0.7.8

# GUI Framework
PySide6>=6.6.1
PyQt6>=6.6.1

# Firebase and Google Cloud
firebase-admin>=6.4.0
google-cloud-firestore>=2.13.1
google-cloud-storage>=2.10.0
google-auth>=2.25.2

# HTTP and API
httpx>=0.25.2
aiohttp>=3.9.1
requests>=2.31.0

# Utilities
loguru>=0.7.2
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.1